development DAO replacing long to int

+++ demo/src/main/java/com/project/marketsimulator/model/produit/Produit.java
@@ -5,10 +5,11 @@ import javax.persistence.*;
 @Entity
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 public abstract class Produit {
+
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false)
-    private long id;
+    private int id;
 
     @Column(nullable = false)
     private String name;
